+++ repos/exposed/src/jvmMain/kotlin/dev/inmo/micro_utils/repos/exposed/ExposedCRUDRepo.kt
@@ -1,6 +1,5 @@
 package dev.inmo.micro_utils.repos.exposed
 
-import dev.inmo.micro_utils.repos.Repo
 import org.jetbrains.exposed.sql.*
 
 interface ExposedCRUDRepo<ObjectType, IdType> : ExposedRepo {